produce-messages/requirements.txt
autopep8==2.3.2
paho-mqtt==2.1.0
pycodestyle==2.14.0
python-dotenv==1.1.1
